Reorganize sandbox
