+++ .env.example
@@ -1,6 +1,9 @@
+# Smart contracts
+STABLE_COIN_CONTRACT_ADDRESS=0x123456789
+
+# Telegram
 TELEGRAM_TOKEN=1234567:asdfghjkl
 TELEGRAM_CHANNEL_ID=@dFusionPoC
 
-#PK=my-private-key
-#MNEMONIC=my super secret mnemonic
+# Ethereum Node
 NODE_URL=http://localhost:8545
